data gets successfully

+++ Swift Coin/Models/Coin.swift
@@ -16,7 +16,7 @@ struct Coin: Codable {
     let totalVolume, high24H, low24H: Double?
     let priceChange24H, priceChangePercentage24H: Double
     let marketCapChange24H, marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
@@ -25,7 +25,7 @@ struct Coin: Codable {
     let sparklineIn7D: SparklineIn7D? 
     let priceChangePercentage24HInCurrency: Double?
     
-    enum CodingKeys: String, CodingKey {
+    enum CodingKeys: String, CodingKey { 
         case id, symbol, name, image
         case currentPrice = "current_price"
         case marketCap = "market_cap"
